Move test.DB commands to superdesk.db
